+++ locals.tf
@@ -4,16 +4,12 @@ locals {
     stack = var.stack
   }
 
-  name_prefix = var.name_prefix != "" ? replace(var.name_prefix, "/[a-z0-9]$/", "$0-") : ""
+  name_prefix  = var.name_prefix != "" ? replace(var.name_prefix, "/[a-z0-9]$/", "$0-") : ""
+  default_name = lower("${local.name_prefix}${var.stack}-${var.client_name}-${var.location_short}-${var.environment}")
 
-  server_name = coalesce(
-    var.server_custom_name,
-    "${local.name_prefix}${var.stack}-${var.client_name}-${var.location_short}-${var.environment}-sql",
-  )
-  elastic_pool_name = coalesce(
-    var.elastic_pool_custom_name,
-    "${local.name_prefix}${var.stack}-${var.client_name}-${var.location_short}-${var.environment}-pool",
-  )
+  server_name = coalesce(var.server_custom_name, "${local.default_name}-sql")
+
+  elastic_pool_name = coalesce(var.elastic_pool_custom_name, "${local.default_name}-pool")
 
   vcore_tiers                 = ["GeneralPurpose", "BusinessCritical"]
   elastic_pool_vcore_family   = "Gen5"