+++ locals.tf
@@ -22,6 +22,12 @@ locals {
     family   = contains(local.vcore_tiers, var.sku.tier) ? local.elastic_pool_vcore_family : null
   }
 
-
   databases_users = var.create_databases_users ? { for db in var.databases_names : db => format("%s_user", replace(db, "-", "_")) } : {}
+
+  allowed_subnets = [
+    for id in var.allowed_subnets_ids : {
+      name      = split("/", id)[10]
+      subnet_id = id
+    }
+  ]
 }