+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Breaking
   * AZ-273: Use for_each to iterate over databases
   * AZ-160: Revamp diagnostic settings
+  * AZ-351: Allow usage of vCore model
 
 Changed
   * AZ-273: Terraform 0.13+ compatible