+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Unreleased
+
+Changed
+  * AZ-398: Force lowercases on default generated name
+
 # v4.0.0 - Unreleased
 
 Breaking